requirements.txt
# Document Processing Pipeline - Dependencies

# LLM Extraction
openai>=1.0.0

# ML Classification
scikit-learn>=1.3.0
joblib>=1.3.0

# Schema Validation
pydantic>=2.0.0

# Web UI
streamlit>=1.28.0

# Data Loading
datasets>=2.14.0

# Environment Variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

